requirements.txt
numpy>=1.24.3
transformers>=4.32.1
pyaudio>=0.2.14
torch
RPi.GPIO